apply StickyGroupedList for chat threads
